--- crawler/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "crawler",
-  "version": "1.0.0",
-  "main": "crawler.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "description": "",
-  "dependencies": {
-    "axios": "^1.7.2",
-    "cheerio": "^1.0.0-rc.12"
-  }
-}
